makefilelar yazıldı

+++ CPP00/ex01/main.cpp
@@ -3,7 +3,7 @@
 
 bool isnumeric(string str)
 {
-	for(int i = 0; i < str.length(); i++)
+	for(size_t i = 0; i < str.length(); i++)
 	{
 		if(!isdigit(str[i]))
 			return false;
